requirements.txt
#django==1.6             # web-framework of choice
MySQL-python     # The Python MySQL bindings
#South>=0.7.3            # Python-Django MySQL migrations
#yolk==0.4.1             # List out installed package versions (on virtualenv)
ipython
django-extensions

# Python service calls
#requests==1.2.3

# System monitoring and Performance
#raven==2.0.3

#git+https://github.com/matteius/parkle.git

# required by jenkins test automation and code coverage reports


# For API
#djangorestframework==2.3.8

# For Gamestate
#redis==2.8.0
#hiredis==0.1.1


